refactor: extract proxy routes + WebSockets into api/routes/proxy.py
Move 4 proxy endpoints (status, alerts, ws/metrics, ws/alerts) plus
ring buffer state and log readers from server.py into routes/proxy.py.
Re-export push_proxy_alert/push_proxy_metrics for backward compat.
Update test patch targets. server.py: 2433 → 2141 lines (-32% cumulative).
